Support stacked borrows in addition to tree borrows
Stacked borrows is a much stricter model,
which I dislike for a number of reasons:

My biggest complaint is the spatial restrictions of stacked borrows,
which heavily restrict the use of interior pointers in favor of pointers to the start of an allocation.
In particular, this means a Brc must be a pointer to a header
rather than pointing directly to the value.
This needlessly slows down Deref to comply with the stricter rules.
In fact, the necessary refactoring to use a `struct BrcInner` type instead of pointer arithmetic
means we need to drop support for the stable channel (see that commit for more details)

Another disadvantage of complying with the stricter stacked borrows rules
is that it forces us to convert several references to raw pointers.