Remove THIS_THREAD_STATE_FLAG access in `do_destroy_shared`.
The function is not always called from the target thread,
so could update the state of the wrong thread.
Treating a thread as dead before it really was could trigger UB.

The access is actually not necessary since the thread-local flag had already been set to death
in the TLS destructor, before the `do_destroy_shared` was even reached.