Increase biased counter from 14 to 20 bits
Using 14 bits was too small for the number of benchmark iterations,
meaning that we were going to the cold path in benchmarks.
As a consequence, the ShortThreadId is limited to 12 bits.
Meaning that only the first 4096 threads can participate in biased RC.

In the future, we should consider moving to 64-bit state words
to increase the allowable counter width and to make room for Weak references.