Avoid invoking drop in decrement_strong
Instead, we return a flag and let the caller execute the drop function.
This means there is only a single call-site for T::drop,
and we can avoid monomorphizing or passing the drop info
to the the decrement_biased functions.

The drop info is still needed for decrement_shared function,
as we might have to add the drop info to the queue.