Always unbias after merging counters, even if dropping
Until now, I didn't see the need to maintain this invariant
if we were going to delalocate the object anyways.

However, if weak references are implemented the way they are in std::sync,
we could "drop" the primary value while weak references keep the header alive.
It is important these weak references see the object as unbiased after being merged,
or else upgrading the weak reference could incorrectly increment the biased count.