Define WeakReal helper type to avoid unwrap
Calling the real() method returns either a WeakReal corresponding to an actual allocation,
or `None` if the `Weak` corresponds to the special reserved values.

This means that a check for the reserved value only needs to happen once.
While the previous checks should've been removed by the optimizer,
this is still a cleaner design.