More From impls, bringing Brc closer to Arc
The conversions for OsStr, Path and CStr require assumptions
about the internal representation.
For Path, OsStr these assumptions are very reasonable and enabled by
default in the 'nightly' feature.

Making this assumption for CStr is less kosher,
as comments in the stdlib announce intention to change it.

Note that use of the `ptr_meta` crate already makes fragile
assumptions about layout of stdlib types.