Forbid deferred destructors from unwinding
This was the case in practice before, but it was not actually documented.
I have decided not to add an option to disable this behavior,
aside from the option to disable implicit collections entirely.
I think that unwinding from implicit collections is a really bad idea.

The `collect` function does allow destructors to unwind.
We add a `collect_nounwind` function to explicitly forbid the behavior.